feat(home): extract HomeContent and use Suspense
Move the client-side home page logic and UI into a new component
at src/components/HomeContent.tsx to reduce duplication and enable
code-splitting. Replace the heavy implementation in
src/app/page.tsx with a lightweight Suspense wrapper that lazy loads
HomeContent.

The behavior is preserved (search, loader, article display, finance
warning and animations) while isolating stateful logic for easier
maintenance and smaller initial bundles.
